strating coding in Javascript
Class AuthenticationManager is intended to manage authentication-related tasks.
Constructor is a  method that is automatically called when an object is created from the class. In this constructor, an instance of the AuthenticationManager class is being initialized.
this.passcodeMap = new Map();: Inside the constructor, a property called passcodeMap is created and assigned a new instance of the Map object. A Map in JavaScript is a collection of key-value pairs, and it's commonly used to store and retrieve data based on keys.
passcodeMap will be used to store passcodes for authentication purposes
AuthenticationManager class with an empty passcodeMap. The next steps would involve adding methods to manage and manipulate passcodes within this class, such as adding, removing, or validating passcodes.