Validating Passcode
validatePasscode that takes a passcode as a parameter

if (this.passcodeMap.has(passcode)) { ... }: checks whether the passcodeMap has an entry for the given passcode. If it does, the code within the block is executed.

const currentTime = new Date().getTime();: This line creates a variable currentTime and assigns it the current timestamp in milliseconds. This timestamp represents the current time when the validation is being performed.

const passcodeData = this.passcodeMap.get(passcode);includes the timestamp when the passcode was generated and its duration.

return currentTime - passcodeData.timestamp < passcodeData.duration; This line calculates the time elapsed since the passcode was generated (currentTime - 
passcodeData.timestamp) and compares it to the passcode's specified duration (passcodeData.duration). If the time elapsed is less than the duration, it means the passcode is still valid, and the method returns true. Otherwise, it returns false.

return false; If the passcode doesn't exist in the passcodeMap, the method returns false, indicating that the passcode is not valid.